+++ src/daf/logging/logger_json.py
@@ -5,7 +5,8 @@ from .tracing import trace
 from ..misc import doc, async_util
 from ..misc.instance_track import track_id
 
-from .logger_base import _limit_dates, LoggerBASE, C_FILE_NAME_FORBIDDEN_CHAR, C_FILE_MAX_SIZE
+from .logger_base import _limit_dates, C_FILE_NAME_FORBIDDEN_CHAR, C_FILE_MAX_SIZE, LoggerBASE
+from .logger_file import LoggerFileBASE
 
 import json
 import pathlib
@@ -19,7 +20,7 @@ __all__ = ("LoggerJSON",)
 
 @track_id
 @doc.doc_category("Logging reference", path="logging")
-class LoggerJSON(LoggerBASE):
+class LoggerJSON(LoggerFileBASE):
     """
     .. versionchanged:: v3.1
         The index of each log is now a snowflake ID.
@@ -53,30 +54,7 @@ class LoggerJSON(LoggerBASE):
         path: str = str(pathlib.Path.home().joinpath("daf/History")),
         fallback: Optional[LoggerBASE] = None
     ) -> None:
-        self.path = path
-        self._sequence_number = 0
-        super().__init__(fallback)
-
-    def initialize(self):
-        trace(f"{type(self).__name__} logs will be saved to {self.path}")
-        return super().initialize()
-
-    def _generate_snowflake(self) -> int:
-        """
-        Generates an unique snowflake index (id) for identifying logs.
-        The returned number is not fixed size and it consists of
-        <sequence number> | <timestamp in ms since epoch>.
-        <sequence number> is of fixed size 8 bits.
-        """
-        stamp = int(time.time() * 1000)
-        seq = self._sequence_number
-        snowflake = (
-            seq |
-            (stamp << 8)
-        )
-
-        self._sequence_number = (seq + 1) % 0xFF  # modules by max value of 8 bits
-        return snowflake
+        super().__init__(path, fallback)
 
     @async_util.with_semaphore("_mutex")
     async def _save_log(
@@ -191,7 +169,7 @@ class LoggerJSON(LoggerBASE):
     ):
         after, before = _limit_dates(after, before)
         logs = []
-        for filename in self._get_files():
+        for filename in self._get_files(".json"):
             with open(filename, 'r', encoding="utf-8") as reader:
                 data = json.load(reader)
 
@@ -229,70 +207,6 @@ class LoggerJSON(LoggerBASE):
             sorted_ = sorted_[:limit]
 
         return sorted_
-
-    async def analytic_get_num_messages(
-        self,
-        guild: Union[int, None] = None,
-        author: Union[int, None] = None,
-        after: datetime = datetime.min,
-        before: datetime = datetime.max,
-        guild_type: Union[Literal["USER", "GUILD"], None] = None,
-        message_type: Union[Literal["TextMESSAGE", "VoiceMESSAGE", "DirectMESSAGE"], None] = None,
-        sort_by: Literal["successful", "failed", "guild_snow", "guild_name", "author_snow", "author_name"] = "successful",
-        sort_by_direction: Literal["asc", "desc"] = "desc",
-        limit: int = 500,
-        group_by: Literal["year", "month", "day"] = "day"
-    ) -> list:
-        logs = await self.analytic_get_message_log(
-            guild,
-            author,
-            after,
-            before,
-            guild_type=guild_type,
-            message_type=message_type,
-            limit=None
-        )
-        cuts = {}
-        if not len(logs):
-            return []
-
-        regions = ["day", "month", "year"]
-        regions_left = list(reversed(regions[regions.index(group_by):]))
-
-        sort_by_values = get_args(LoggerJSON.analytic_get_num_messages.__annotations__["sort_by"])
-
-        def get_region_text(stamp):
-            return '-'.join(f"{getattr(stamp, region):02d}" for region in regions_left)
-
-        for log in logs:
-            time_group = get_region_text(log["timestamp"])
-            guild = log["guild"]
-            author = log["author"]
-            group_value = time_group, guild["id"], author["id"]
-            if group_value not in cuts:
-                cut_group = cuts[group_value] = [
-                    time_group,
-                    0,
-                    0,
-                    guild["id"],
-                    guild["name"],
-                    author["id"],
-                    author["name"]
-                ]
-            else:
-                cut_group = cuts[group_value]
-
-            if log["success_rate"] == 100:
-                cut_group[1] += 1
-            else:
-                cut_group[2] += 1
-
-        # key: first index is timestamp group, so offset by one and then calculate index by annotation position
-        return sorted(
-            cuts.values(),
-            key=lambda row: row[sort_by_values.index(sort_by) + 1],
-            reverse=sort_by_direction == "desc"
-        )[:limit]
     
     async def analytic_get_invite_log(
         self,
@@ -306,7 +220,7 @@ class LoggerJSON(LoggerBASE):
     ) -> list:
         after, before = _limit_dates(after, before)
         logs = []
-        for filename in self._get_files():
+        for filename in self._get_files(".json"):
             with open(filename, 'r', encoding="utf-8") as reader:
                 data = json.load(reader)